Remove turndown for the etcd node
Subject to further investigation. Currently remove this
line as a temporary fix. It doesn't not affect the current
etcd creation. This turndown is needed to do inplace
replacement of etcd. The turndown ensures that old etcd
instances are removed before new ones are created.
Similar changes are done for the master/worker node.